Game Engine Skeleton and Simple Demo Assets
I created the classes necessary for the game engine, and I also created
a couple assets to test with.
The game engine classes need to have the methods filled out. The main
function also needs to be written.
After writing all of this I noticed there were some errors with our
design, so follow the method interfaces that are here. The major things:
Interactable is now a class because otherwise the name String would be
static for all Interactable objects.
the do() function is now interact(String action, Interactable target)
because do is a keyword, and the parser should identify the target, not
the item.
The Map class is now the World class, and holds important Aspects and
Items as well as all of the Locations. Not all Aspects will need to be
stored in here, for example, a North aspect is simple enough that it can
be created and stored in its Location.
The demo assets basically start you off in a prison cell. the key is on
the floor. you unlock the door and you can go out into the hallway.