Finished Demo Assets and added some more methods
The Demo assets are now done. I also added some more empty methods we'll
need.
One bug is that when we construct all of our locations, the passageways
reference other locations that haven't been built yet. So we'll need to
figure out something to do. We could have 2 build phases: One where we
initialize everything and then in the second phase we link it all
together. It would be really messy though.
I wonder if we could have all the static locations start out as null and
then construct everything...