Parser Completed
I tested it for various inputs and it can correctly identify what the
object, action, and target are. However, I was running into some
problems with searching for the interactables because the names of the
key and door was null.
I also added the parser function for the menu to the Parser Class, so
you can use that in the main menu.
The parseInputAndExecute() function returns false if it couldn't
understand the input, so try asking again if it returns false.