game loop, bug fixes
Finished main game loop, integrating parser and an aspect queue. Also
fixed a few parser bugs that were causing errors. Changed the pickUp()
method in player to pick up an object that contains the keyword if it is
the only match. Added a globalAspects List to World. Various other
fixes.